Remove excessive debugging.
A lot of this was code I added mainly for when I was originally developing this
to make sure it was behaving correctly.  I'm not convinced it's of much use to
users, and added the overhead of extra sub calls for every template token, so
I've decided to just remove it.